requirements.txt
strands-agents>=1.0.0
mcp-server>=0.1.0
boto3>=1.28.0
requests>=2.31.0
aiohttp>=3.8.0
flask>=2.0.0
python-dotenv>=0.19.0
ngrok>=0.8.1